finished refactoring experiment-builder and model-testing
